[REFACTOR] Make a separate service for dynanamic Catalog configuration
**RATIONALE**
The `SolrService` has grown to contain two relatively different
sets of functionality over time:
1. Managing and persisting the connection to a Solr host
2. Translating `Field` settings into the desired CatalogController
   configuration

This change extracts the CatalogController related logic to a
separate service. This change is made in anticipation of extending
the configuration logic to add support for search field configuration.
