+++ app/models/solr_service.rb
@@ -11,7 +11,7 @@ class SolrService < ApplicationRecord
 
   before_create :check_for_existing
   before_destroy :check_for_existing
-  after_commit :update_catalog_controller
+  after_commit { CatalogConfigService.update_catalog_controller }
 
   def self.current
     SolrService.first || SolrService.create(DEFAULT_CONFIG)
@@ -72,45 +72,11 @@ class SolrService < ApplicationRecord
     true
   end
 
-  def update_catalog_controller # rubocop:disable Metrics/AbcSize
-    CatalogController.configure_blacklight do |config|
-      config.connection_config[:url] = solr_connection_from_config
-      config.facet_fields = blacklight_fields_from_config.facet_fields
-      config.index_fields = blacklight_fields_from_config.index_fields
-      config.show_fields = blacklight_fields_from_config.show_fields
-      config.index.title_field = title_field_from_config
-    end
-  end
-
   def self.solr_connection
     CatalogController.blacklight_config.repository.connection
   end
 
   private
-
-  def blacklight_fields_from_config
-    config = Blacklight::Configuration.new
-    # NOTE: skip the first field because it's always used as the main title field for Blacklight
-    Field.active_in_sequence[1..]&.each do |field|
-      update_field(config, field)
-    end
-    config.add_show_field 'files_ssm', label: 'Files', helper_method: :file_links
-    config
-  end
-
-  def update_field(config, field)
-    config.add_facet_field(field.solr_facet_field, label: field.name, limit: 10) if field.facetable
-    config.add_index_field(field.solr_field_name, label: field.name) if field.list_view
-    config.add_show_field(field.solr_field_name, label: field.name) if field.item_view
-  end
-
-  def title_field_from_config
-    Field.active_in_sequence.first&.solr_field_name
-  end
-
-  def solr_connection_from_config
-    "#{solr_host}/solr/#{solr_core}"
-  end
 
   def check_for_existing
     raise ActiveRecord::RecordInvalid if SolrService.count >= 1
